Agent-operator MCP surface over benchmark sidecars (understudy benchmarks mcp)
Lift the hub's pure loader/validation logic into the CLI package
(benchmark-hub-types/-core, benchmark-replay) with the app re-importing
from dist — extending the runs-core anti-drift pattern — and serve
list/read/diff/review/queue tools over stdio for coding agents. The
server never executes models; queue_run only writes
understudy.run_request.v1 files for the executor/daemon.
