+++ src/commands/benchmarks.ts
@@ -1,0 +1,31 @@
+import { Command } from "commander";
+
+/**
+ * `understudy benchmarks …` — the agent-operator surface over the file-based
+ * benchmark artifacts. `benchmarks mcp` serves the read/diff/review/queue
+ * tools over stdio so a coding agent can drive the improvement loop; it
+ * never executes models (that stays with `understudy runs execute`).
+ */
+export function registerBenchmarksCommand(program: Command): void {
+  const benchmarks = program
+    .command("benchmarks")
+    .description("Agent-operator surface over local benchmark dirs (proposed foundry outputs + promoted benchmarks)");
+
+  benchmarks
+    .command("mcp")
+    .description(
+      "Stdio MCP server: list/read benchmarks and tasks, read + diff rollout trajectories with contract " +
+        "scoring, append reviews, queue run requests, and poll run status — same shared cores as the hub, " +
+        "no model execution",
+    )
+    .option(
+      "--root <dir>",
+      "Additional benchmark root scanned after ~/.understudy/benchmarks (repeatable)",
+      (value: string, previous: string[]) => [...previous, value],
+      [] as string[],
+    )
+    .action(async (options: { root: string[] }) => {
+      const { runBenchmarksMcpServer } = await import("../benchmarks-mcp.js");
+      await runBenchmarksMcpServer(options.root);
+    });
+}
